% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_atbash.R
\name{as_atbash}
\alias{as_atbash}
\title{Convert Hebrew text to AtBash form.}
\usage{
as_atbash(text)
}
\arguments{
\item{text}{Hebrew Text}
}
\description{
Get AtBash configuration of of Hebrew Text
}
\examples{
tct<-"דוד המלך"
as_atbash(tct)
}
\concept{AtBash}
\keyword{Hebrew}
