% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_gemmatria.R
\name{as_gemmatria}
\alias{as_gemmatria}
\title{Convert to Gemmatria Form}
\usage{
as_gemmatria(number)
}
\arguments{
\item{number}{a positive integer}
}
\description{
This converts positive integers into gemmatria form
}
\examples{
nb<-613
as_gemmatria(nb)
}
\concept{Gemmatria}
\keyword{Hebrew}
