% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gemmatria.R
\name{get_gemmatria}
\alias{get_gemmatria}
\title{Get Gemmatria of Text}
\usage{
get_gemmatria(text)
}
\arguments{
\item{text}{Hebrew text}
}
\description{
Get the gemmatria of Hebrew words.
}
\examples{
# heb<-"שלום"


# get_gemmatria(heb)

# get_gemmatria_st(tct)

}
\concept{Gemmatria}
\keyword{Hebrew}
