% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gemmatria_atbash.R
\name{get_gemmatria_atbash}
\alias{get_gemmatria_atbash}
\title{Get AtBash Gemmatria of Text}
\usage{
get_gemmatria_atbash(text)
}
\arguments{
\item{text}{Hebrew text}
}
\description{
Get the AtBash Gemmatria of Hebrew text
}
\examples{

heb<-"שלום"
get_gemmatria_atbash(heb)
}
\concept{AtBash}
\concept{Gemmatria}
\keyword{Hebrew}
