% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gemmatria_rt.R
\name{get_gemmatria_rt}
\alias{get_gemmatria_rt}
\title{Get Gemmatria of Roshei Taivos}
\usage{
get_gemmatria_rt(text)
}
\arguments{
\item{text}{Hebrew text}
}
\description{
Get the Gemmatria of the Roshei Taivos of Hebrew text.
}
\examples{

tct<-"דוד המלך"
get_gemmatria_rt(tct)
}
\concept{Gemmatria}
\concept{Roshei Taivos}
\keyword{Hebrew}
