% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gemmatria_st.R
\name{get_gemmatria_st}
\alias{get_gemmatria_st}
\title{Get Gemmatria of Sofei Taivos}
\usage{
get_gemmatria_st(text)
}
\arguments{
\item{text}{Hebrew text}
}
\description{
Get the Gemmatria of the Sofei Taivos of Hebrew text
}
\examples{

tct<-"דוד המלך"
get_gemmatria_st(tct)
}
\concept{Gemmatria}
\concept{Sofei Taivos}
\keyword{Hebrew}
