% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_gemmatria.R
\name{as_gemmatria}
\alias{as_gemmatria}
\title{Convert to Gemmatria Form}
\usage{
as_gemmatria(number)
}
\arguments{
\item{number}{a whole number}
}
\description{
This converts decimal (base ten) _whole_ numbers into gemmatria form
}
\concept{Gemmatria}
\keyword{Hebrew}
